Update lottery contract.

+++ chain/contract/AElf.Contracts.LotteryContract/LotteryContractConstants.cs
@@ -5,6 +5,6 @@ namespace AElf.Contracts.LotteryContract
         private const long DefaultPrice = 1;
         private const long DefaultDrawingLag = 80;
         private const long MaximumBuyAmount = 100;
-        private const long MaximumReturnAmount = 20;
+        private const int MaximumReturnAmount = 20;
     }
 }
